feat : category 이름 가려지는 이슈 수정
