+++ index.html
@@ -47,7 +47,7 @@
         </div>
     </main>
 
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="./category.js"></script>
 
 </body>
